fix: remove erroneous lifetime
Whilst first learning Rust, I did not appreciate the implicit lifetime
created when taking `self` by reference in a method. I misinterpreted
an error message and created an explicit lifetime which was thenceforth
propagated up the trait hierarchy despite being meaningless. Because
the shared iterator implementation for arrays relies upon unsafe
pointer manipulation whereby the programmer (me) promises the derived
lifetimes are correct, everything 'just worked' until now.

This commit changes the signatures of `Linear` methods to have the form
`fn iter<'new>(&'new self) -> impl Iterator<Item = &'new T>` thus
connecting the lifetime of elements obtained via an iterator to the
lifetime of that iterator and ultimately back to the lifetime of the
object the iterator was created from.

If I now understand correctly, prior to this commit, the yielded
elements have a lifetime associated with the object the iterator was
create from, but _NOT_ the iterator itself hence the iterator could
be dropped while references to elements were still accessible.