revert: external node count
Having an external node count is unnecessary for a linked list and
goes against the ethos necessitating that everything which modifies the
list must also modify the count independently merely increasing the
area for bugs to develop.

Although having such a count would allow iterating from tail to head
when that chain is shorted to the requested node, it complicates the
logic of all operations without altering the asymptotic
time/memory/resource complexity of those operations.

Instead, custom implementation for prepend/append and the like should
be implemented with explicit performance guarantees that are obvious
to the caller instead of asking them to trust our implementation will
make the optimizations they are expecting.