fix: only do pointer arithmetic when necessary
In the case where there was an allocation, the derived pointer is valid
even when it points one byte past the end of the allocated object, and
`ptr::copy` would only read from that pointer if there were in fact
trailing elements to copy. However, in the case where there is no
allocation, the underlying pointer being derived from is dangling so it
is _NOT_ valid to increment it.