refactor: loop refactoring circumvents continue
Previous refactoring of the loop as a whole means the algorithm no
longer needs to explicitly continue under those circumstances, simply
allowing control flow to continue with existing scope will update for
the next iteration.

Moreover, it would not be valid to break from the loop at the point as
conjectured because balancing the immediate parent does not preclude
imbalancing further ancestors.