revert: remove panic condition
This panic condition was originally added to help with the insertion
and removal logic for the encompassing AVL tree wherein it would be a
logic error to invoke a rotation when there is not an imbalance.
However, that is a recoverable error which does not invalidate AVL
invariants hence it should not result in a panic. With test
infrastructure now existing, such errors within callers of rotate
left/right will be detected without panicking. Therefore, these methods
should not consider such conditions unreachable instead leaving the
caller in control.