test: into iterator drops when both ends advanced
This commit updates the tests to handle all combinations of front and
back advances instead of merely equivalent amounts.

Alongside that, this commit also modifies the styling of
`drops_unyielded_elements_when_advanced_from_front`
`drops_unyielded_elements_when_advanced_from_back`
to be consistent with the improved comment explanations in
`drops_unyielded_elements_when_advanced_from_both_ends`.

Moreover, as the final change this commit makes (sorry this should have
been split but it was awfully connivent to do it in one pass or else
risk divergences) is to reduce the size of the input arrays to only
eight (8) elements. This is useful for both-ends tests because they
have time complexity of O(N<sup>2</sup>), and the manually declared
arrays are only six (6) elements so it is enough to cover all possible
partitions and can be increased on a per-test basis if I am wrong.