doc: remove reference to immutable memory
Dope is inherently mutable and specific to accessing mutable memory.
Moreover, construction of the object necessitates a `NonNull` which
itself necessitates a mutable pointer. Since using Dope atop immutable
memory would require casting an immutable pointer to mutable, the
caller is already taking on the responsibility and we should not
erroneously imply such behaviour is supported.