fix: correctly index leaves of some root
Previously, sift-down operations were restricted via a subslice to
exclude parents of the current root of the relevant subtree. However,
this caused indexing failures because the functions for left and right
children would wrongly produce elements from other subtrees since they
are interlaced into the underlying array.

For example, consider the 5th element in the overall array. The correct
left child is 2 * 5 + 1 = 11. However, if as before this commit it is
presented as the first element of a subslice, then its relative left
child is 2 * 0 + 1 = 1, which is index 6 in the overall array. Needless
to say, 11 != 6.