chore: add file blacklist for git
Git automatically tracks when new files have been added to the
repository which can be a nuisance in some circumstances because there
exists local artifacts which should not be pushed into remove version
control.

This commit adds a file which will disable that automatic tracking
for files which are almost certain to exist as a consequence of
development but should not be tracked by git.